updated migration script!
migration is working now, but slow -> took about 4 hours to update full size openuss db.

git-svn-id: https://svn.code.sf.net/p/openuss/code/trunk@3377 60046477-0c10-0410-bf4b-e552a4ff1768

+++ openuss-plexus/plexus/plexus-core/src/main/java/org/openuss/migration/MigrationServiceImpl.java
@@ -126,8 +126,8 @@ public class MigrationServiceImpl
     			User user = getCourseMemberDao().courseMemberInfoToEntity(member).getUser();
 
     			Permission permission = getSecurityService().getPermissions(user, course);
-    			if (permission.getMask().intValue()==1040){
-    				getSecurityService().removePermission(user, course);
+    			if (permission!=null&&permission.getMask().intValue()==1040){
+   					getSecurityService().removePermission(user, course);
     			}
     			getSecurityService().addAuthorityToGroup(user, participantsGroup);
     		}
